sodiumdecrypter: Ensure to pull enough data from upstream if not starting on a chunk boundary
We round down to the previous chunk boundary, but then we also have to
add the difference between the offset of that chunk to the requested
offset to the size of the buffer that is pulled from upstream. Otherwise
there will be an unneeded short read, which could cause downstream to
EOS too early.