ci: have build jobs depend on either build-* or update-* jobs
followup to 23c07d3c

When we trigger an update pipeline, instead of build-jobs we
replace them update-* jobs. In order for gitlab to not complain
add both of the jobs in the needs array, but mark them as optional
so either of them can be used depending on the usecase.